Inline `utils/helpers.js` file
